Run DataGenerator to create data of N size and specify location.
Run JoinRunner to evaluate 2 different types of joins.
The naiveJoin will require a data shuffle and does nothing to address the obvious data skewness for records from the US.

The bucketJoin function addresses this concern by bucketing the data into N buckets. Thus avoiding a shuffle altogether as shown in the physical plans. This is assuming that the incoming data will already be bucketed as bucketing is a wide operation and will require a shuffle. This approach is best for datasets that will be used many times.

On real data the joining of 2 bucketed datasets would be significantly faster than a regular SortMergeJoin. The join could further be sped up by partitioning the data, but we would have to define some salting logic so that all the records with COUNTRY=US don't end up in the same partition.

In Props.scala you can see all the inputs needed to run JoinRunner. These are mixed in as a trait to keep things clean. These inputs should be passed as env vars such as `set bucket_num=100`